refactor(sight): extract background threads module
Move the three background-thread spawners (config watcher,
token-collector watcher, stale-pending scanner) and their two config
helpers out of unified.rs into a new background.rs module, and factor
the repeated stop-signal sleep loop into utils::thread::sleep_or_stop.

Pure relocation: thread bodies, log messages, timing (sleep-first),
and the uid-check process::exit are preserved verbatim. unified.rs
drops from 2263 to 1695 lines (now under the 2000-line hard rule).

Extracting the loop bodies into free functions makes them unit-testable
without BPF: background.rs and utils/thread.rs carry the migrated tests
plus new coverage for sleep_or_stop interval semantics and the
stale-scanner marking a seeded pending row interrupted.
